Fix draw point VAO not being initialized
Former-commit-id: bd9a0040f0dd22fe9258757e486cb917b5518a1c

+++ Mountain/inline/Mountain/utils/utils.inl
@@ -110,7 +110,7 @@ namespace Mountain
     template <uint64_t Offset, uint64_t Count>
     constexpr uint64_t Utils::GetBits(const uint64_t value)
     {
-        constexpr uint64_t mask = (static_cast<uint64_t>(1) << Count) - 1;
+        constexpr uint64_t mask = (1ull << Count) - 1;
         return value >> Offset & mask;
     }
 }
